+++ src/main/java/controller/Controller.java
@@ -1,7 +1,7 @@
 package controller;
 
-import io.request.HttpRequest;
-import io.response.HttpResponse;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 
 public interface Controller {
     void handle(HttpRequest request, HttpResponse response);
