+++ src/main/java/filesystem/FindResult.java
@@ -1,4 +1,4 @@
-package io.response;
+package filesystem;
 
 import enums.ContentType;
 import enums.Status;
