+++ src/main/java/http/common/Body.java
@@ -1,4 +1,4 @@
-package io.response;
+package http.common;
 
 public class Body {
