+++ src/main/java/http/common/Protocol.java
@@ -1,4 +1,4 @@
-package io.request.startLine;
+package http.common;
 
 import java.util.Arrays;
