+++ src/main/java/http/request/HttpRequest.java
@@ -1,18 +1,16 @@
-package io.request;
-
-import io.request.startLine.StartLine;
+package http.request;
 
 import java.util.HashMap;
 import java.util.Map;
 
 public class HttpRequest {
 
-    private StartLine startLine;
+    private RequestStartLine startLine;
     private Header header;
 
     private RequestParameter requestParameter;
 
-    public HttpRequest(StartLine startLine, Header header, RequestParameter requestParameter) {
+    public HttpRequest(RequestStartLine startLine, Header header, RequestParameter requestParameter) {
         this.startLine = startLine;
         this.header = header;
         this.requestParameter = requestParameter;
