+++ src/main/java/http/response/enums/ResponseAttribute.java
@@ -1,4 +1,4 @@
-package io.response;
+package http.response.enums;
 
 public enum ResponseAttribute {
     CONTENT_TYPE("Content-type"),
